requirements.txt
fastapi[standard]>=0.128.5
uvicorn[standard]>=0.40.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
neo4j>=5.28.0
pydantic-settings>=2.0.0
